multi: opt for plain http connections.
this removes tls infrastructure from the mining pool
and the client in favour of using http connections for
message delivery.  This is as a result of a discussion
I had with davecgh where he pointed out that most
miners are not happy with using tls, they claim the
overhead puts them at a disadvantage with regards
to speed.

This also better handles websocket disconnections.