multi: track active websocket connections.
This adds an active websocket
connection counter and leverages
that to in determining when notifications
should be broadcast, ie. when there are
connected clients.

Only requests can be notification currently
as per the  to  the json 2.0 spec.

fetchBlockHeight, fetchNonce, fetchID and their
corresponding setters have been added for parsing
work data.