multi: use higher share gen. time in solo pool mode.
This updates the pool diff helper to use 25 seconds as
the max gen. time when calculating the appropriate
diffs for pool clients in solo pool mode where valid shares
are a non-factor. The max gen. time was was kept under
30 seconds due to the likelihood of pool clients having
timeouts of 30 seconds when there is no activity. This
also updates the hashrates of the DR5 and D1.