multi:  add mined block records.
This adds mined block records which are detailed
records of a block mined by the pool which allow
for auditing, client lookup and reporting
functionality for the pool.

This also simplifies the updates accepted work
struct to be persisted as mined work. All unused
functions and fields particularly for a stratum
client have been commented out.

This also reverts the exclusion of simnet from possible
mining networks to choose from,  particularly to allow testing
asics on  simnet.