multi: update work notification pipeline.
This updates the work notification pipeline due
to a bug that was leaving connected miners on
stale work each update cycle. This also improves
log messages outputted and reverts an update
that broke json-rpc compatibility for notifications.

This also removes the username requirement in
order to be authorized by a pool when solo pool
mode.