multi: update mined work tracking.
This reworks the mined work tracking of the pool. Accepted work
when confirmed as mined are marked as such. Processed shares
are now pruned after generating payments from them.

An endpoint has been added to the api for database backups.
The readme has been updated with the new endpoint as well.