multi: implement realtime pool hash rate.
This tracks the hashrate of pool clients and based on the
intervals between their work submissions. This also fixes a
bug with the cpu miner shutdown process where terminated
 goroutines were not signalling the waitgroup of being done.

The cpu miner connection handler is also less aggressive in
reconnecting  to the pool when it drops, it now waits 5
seconds for every retry.