network: fix accepted work confirmation.
This fixes the accepted work confirmation bug by
fetching the parent accepted work of the connected
block instead of fetching the accepted work of the
connected block first. This is being done because
it is possible for the accepted work of the
connected block to be unavailable at the time its
corresponding block connected notification is
recieved. It's also more efficient to directly fetch
the parent accepted work since that's ultimately
the accepted work being confirmed.