pool: update payDividends.
This updates the payment manager's payDividends method to
immediately make payments laying claim to the spendable
coinbase. The updated payout  transaction does not return
any change since the majority of it is distributed to associated
accounts of payments, the rest as pool fees. Th payout transaction fees are deducted from the pool fees
collected from the spendable coinbase.

Tests have been added for the new methods of the payment
manager.