pool: filter out dust payments.
This updates payment generation to filter out
dust payments. These dust payments will be
forfeited by the accounts generating them
and be added to the pool fee payout. This is
being done since dust outputs will cause the
payout transaction to error and also serve as
a deterrent to miners sporadically contributing
to a pool.