Handle database IDs as strings where possible.
Regular conversion of IDs between string and []byte is necessary because bbolt
expects all keys and values to be byte arrays. Because this is specific to
bbolt and we want to generalise the code to work with other database
implementations, the majority of our code should handle IDs as strings and only
convert to []byte as close to the bbolt code as possible.

This has a nice side-effict of removing a large amount of inline type
conversion.