Remove mode switching behaviour. (#269)
This removes the behaviour which would automatically empty the database and create a new one when the pool mode was changed in config. The new behaviour is to log an error and prevent the pool from starting up.

The reason for this is the empty/backup behaviour does not map well onto other database implementations such as postgres, and after discussing with dnldd we agreed that its better for an admin to explicitly make the decision to backup and create a new DB, rather than the code doing it.

Also added test coverage for the poolmode storage/retrieval from database.