pool: Keep hub waitgroup local.
The waitgroup for the hub is currently being passed down into all
various configuration structs for each of the goroutines it runs which
requires a lot of plumbing and is quite error prone.

This addresses that by modifying the logic to keep the hub waitgroup
local to the run method and removes the aforementioned unnecessary
plumbing.

These changes have also exposed an issue with endpoint.run in that it
never terminates, but this commit does not fix that issue.