pool: Gen extra nonce before initialize.
This modifies the new client logic to create the extra nonce before to
creating the client and using it during initialization.  Constructors
should generally handle everything that can error prior to creating the
object since it is wasteful to create an object only to immediately
throw it away when there is an error.