multi: No string pointer on submit work.
This modifies the SubmitWork function to take the submission directly
instead of a pointer to it.

Strings are already reference types, so there is no reason to take a
pointer to a string unless the calee intended to change what the
caller's string points at. Moreover, taking a pointer could also lead to
unnecessary allocations.