harness: Make the harness more Windows friendly.
Msys automatically transforms paths for executed commands, but it is
unable to do that for paths in the config files.  This updates all of
the paths in the config files to use relative paths instead so they work
with the various processes.

It also use the posix compliant shebang for locating bash, adds a new
variable to control whether or not the profiling ports are enabled, and
disables postgres by default.