multi: Make GUI run blocking and wait for shutdown.
This updates the gui.Run method to block until the context is canceled
as expected and then adds a waitgroup across both it and the hub to
allow for clean shutdown of both subsystems prior to shutdown.

It also moves the initial connection and work acquisition prior to
launching the subsystems to avoid launching them at all in the event of
an error that would prevent the hub from initializing.