gui: Init cache when GUI is created.
This reworks the GUI cache initialization to happen when the GUI is
created.

It accomplishes this by updating the init func to accept the config
instead of individual components and moving the logic that acquires the
details via the config from the Run method into it.

It also unexports the InitCache method since it is only ever used
internal to the gui package.

Not only does it further consolidate the creation logic in one place, it
also exposes any errors back to the main thread when the pool is first
initializing so any issues can be corrected immediately.