gui: Split out cache and websocket update run logic.
This moves the code that periodically updates the cache and notifies
websocket clients into a separate method that is invoked by run.

It also runs the updates in the background and introduces a waitgroup
for it.  The intention is to have the webserver covered by the waitgroup
in a future commit.