pool: Cleanup client read error logging.
This switches the code that examines read errors for timeout to use the
proper interface from the net package that is designed for that purpose
instead of digging into concrete types.

It also ensures clients that timeout without ever identifying themselves
have their uniquely-assigned extra nonce used as the id instead so the
log messages can be attributed to the ip of the offending clients.

Finally, it no longer logs an error when an io.EOF is encountered as
that is an expected condition when the client disconnects and is not an
actual error.