pool: Remove client shutdown callbacks.
This reworks the client shutdown interaction with the endpoint so that
it's handled directly by the endpoint versus passing callbacks down to
the client which the client must then invoke.

This approach is more robust since it guarantees correctness without
relying on the client invoking callbacks at the appropriate times which
is something that could easily break with a code refactor.

Moreover, this approach is also easier to reason about for correctness
since it consolidates the logic that adds and removes the client from
the endpoint directly at the call site within the endpoint itself as
compared to needing to trace through the entire lifecycle of a client to
ensure it's actually invoking the proper callbacks.