pool: Leave failed postgres db state in pool tests.
This modifies the pool tests to avoid cleaning up the postgres data when
there is a test failure and the -failfast test flag is set so the
contents of the database can be examined.

Since a failed run with -failfast will no longer purge the data, it also
adds logic to purge the data prior to running the tests to ensure any
previous failed runs are cleaned up on subsequent runs.