pool: Improve share UUID with random value.
The postgres tests currently fail when running on fast systems due to
the lack of a generating unique share IDs.

After digging into the failure a bit, the root cause is that the
supposed UUID used for the share id is not created with any randomness
to it.

As the test failure demonstrates, it is generally not a good idea to
rely solely on timestamps for generating supposedly unique values.
While it is fairly unlikely that two shares will come in at the same
nanosecond in most cases, there are a variety of factors that can lead
to such a situation.

So, this appends a random value to the share UUID in order to help
ensure a unique overall value and updates the callers accordingly.  It
also updates the database upgrade code to ensure the old share id format
is used when performing the old version upgrade.

A new database upgrade is not needed because any existing entries are
necessarily already unique or they would've hit duplicate constraint
failures and there is not any code that explicitly relies on the key
being a specific length beyond the first 8 bytes being the timestamp.

Ideally, a v4 UUID would be created according to RFC 4122 instead,
however, that would be a much more invasive change since the current
bolt database code uses the UUID as a key and relies on the timestamp
being the first 8 bytes in the key for filtering purposes.