pool: Improve payment UUID with random value.
The UUID used for the payment ID was not created with any randomness to
it, leading to occasional test failures due to duplicated values.
Appending a random value to the UUID helps to ensure uniqueness.

The commit also updates the database upgrade code to ensure the old
payment ID format is used when performing the old version upgrade.