pool: Don't potentially leak paymgr goroutine.
This ensures the goroutine in the payment manager that handles sending
rescan responses respects the context so it does not potentially leak
the goroutine if the context is canceled prior to receiving a response.

It also corrects the channel close logic by moving into the goroutine.
As mentioned in previous commits, receivers should never close channels,
only senders once they are sure there are no more goroutines that can
write to the channel.