pool: Correct racy paymgr processing logic.
This corrects the mutex logic in the payment manager that keeps track of
whatever or not payments are already being processed.

In particular, the code that is being updated previously checked and set
the processing flag under two separate locks of the mutex which could
easily result in the first check returning false for two concurrent
goroutines followed both both setting it to true and executing which
would defeat the purpose of the check.