pool: Avoid potential send to closed chan in hub.
This avoids a potential case in the hub that could result in a panic
during shutdown due to sending to a nil channel.

Since the rpcclient callbacks can be invoked at any, even after the
context is done, some additional synchronization would be required to
prevent attempting to send to the closed channels if they are closed
when the context is done.

Rather than adding more elaborate synchronization logic, this simply
chooses not to close the channels to resolve the situation because there
really isn't any reason to close them and not closing them also resolves
a secondary issue.  Namely, the receiving side doesn't check if the
channels are closed, so closing them could actually lead to the channel
being incorrectly selected without an accompanying message which itself
could lead to a panic.

Finally, the receiving goroutine will be exited when the context
completes meaning there won't be anything listening to the channels in
very short order either which is even more reason not to worry about
closing them.