pool: Correct some test failure paths.
The current tests in the client message handling incorrect reference the
fields of a response message in the case the message type isn't the
expected type which would be a panic since the instance variable would
be nil in that case.

This corrects that by showing the expected and received message types
instead.