pool: Rename cpu work handler to clarify default.
This renames the handler that sends work to a miner from handleCPUWork
to sendWorkDefault and accepts the name of the miner as a parameter.

This is intended to help make it clear that it is sending the work and
that it is the default format that all miners should ideally follow.

It also improves a couple of error messages while here.