pool: Do not populate/use gentx2.
This modifies the pool to report the entire partial header via the first
generation tx field as intended and no longer populates the second
generation tx field since it does not apply to Decred.

Some of the old ASICs hacked the stake version in there, but that is
really not something that should have ever been done because the stratum
"protocol" (which is not actually very well defined) already does not
individually provide all of the information the Decred header
needs nor does it provide an official way to extend it.

Further, the Decred header explicitly provides additional space which
removes the need to create a new coinbase and update the merkle root.

So, in order to address these things, the field that was intended to
serve for the coinbase (generate transaction) was repurposed to contain
the serialized partial header for data after the previous block hash in
the format it is to be hashed.  Therefore, it should be providing the
entire remaining partial header to ensure that any future modifications
to the end of the header are available to miners without modification.

Given that the old ASICs no longer work with the network, this takes the
opportunity to make the change.