pool: Remove handlePayments goroutine.
Previously a long-lived goroutine named handlePayments would buffer
calls to processPayments and run them sequentially. This adds
unnecessary complexity without any obvious benefit.

In addition, it appears as though the only purpose of the done channel
in the paymentMsg was to facilitate testing, which means this change
makes it redundant and it can be removed.