pool: Remove output total param from applyTxFees.
The applyTxFees func previously took in a set of outputs as one param,
and the total value of those outputs as a separate param. This
introduced an opportunity for the function to be misued in that it could
be passed a set of outputs and a total which do not match.

An instance of this mistake actually existed in the tests for
applyTxFees. The value of the inputs and outputs was 100 DCR, but the
provided total was 5 DCR.