gui: Display miner ip:port on account page.
The miner ip:port are already present in the connected client info so
they can be displayed on the account page with only a minor change.

Previously only the name of the miner software and the hashrate were
displayed which made it impossible to distinguish which client was
which. Having the ip:port of the client isn't much better, but it's
definitely better than nothing.